+++ test/test_generate/test-autograder/autograder-r-requirements-correct/setup.sh
@@ -28,8 +28,8 @@ if [ "${BASE_IMAGE}" != "ucbdsinfra/otter-grader" ]; then
 fi
 
 # install dependencies with conda
-conda config --set channel_priority strict
 conda env create -f /autograder/source/environment.yml
+conda run -n otter-env Rscript /autograder/source/requirements.r
 
 # set conda shell
 conda init --all