+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+**Unreleased (beta):**
+
+* Converted all dependency management in the autograder configuration zip file to Conda per [#501](https://github.com/ucbds-infra/otter-grader/issues/501)
+
 **Unreleased:**
 
 * Display emojis with test results to more clearly show which tests pass and which fail, as per [#533](https://github.com/ucbds-infra/otter-grader/pull/533).